fix(auth): move ProviderLogin out of the auth page module
A page file may only export the fields Next.js recognizes, so the
test-only `export const ProviderLogin` made `next build --webpack` fail
type checking with "ProviderLogin is not a valid Page export field".
That is the OpenNext Cloudflare deploy path; the turbopack build used by
`pnpm build-web` does not run the page export check, so the break was
invisible locally.

Move the component to src/app/auth/components/ProviderLogin.tsx and
import it from both the page and its test.
